+++ changelog.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## Upcoming
+* Added Explosive Barrels
+  * Highly unstable will explode instantly
 * Out of season crops no longer break. They just don't grow
 * Creepers now drop 1-3 gunpowder instead of 0-2 and have a rare TNT drop
 * More advancements work